Fix/sf employer hierarchy

+++ db/migrations/development/20190529150447_submission_table_edit.js
@@ -1,17 +1,27 @@
 exports.up = function(knex, Promise) {
   return Promise.all([
-    knex.schema.table("submissions", function(table) {
-      table.renameColumn("503_cba_app_date", "seiu503_cba_app_date");
-      table.uuid("submission_id").primary();
+    knex.schema.hasTable("submissions").then(function(exists) {
+      if (exists) {
+        knex.schema.table("submissions", function(table) {
+          table.renameColumn("503_cba_app_date", "seiu503_cba_app_date");
+          table.uuid("submission_id").primary();
+        });
+      }
     })
   ]);
 };
 
 exports.down = function(knex, Promise) {
   return Promise.all([
-    knex.schema.table("submissions", function(table) {
-      table.renameColumn("seiu503_cba_app_date", "503_cba_app_date");
-      table.dropColumn("submission_id");
+    knex.schema.hasTable("submissions").then(function(exists) {
+      if (exists) {
+        knex.schema.table("submissions", function(table) {
+          if (table) {
+            table.renameColumn("seiu503_cba_app_date", "503_cba_app_date");
+            table.dropColumn("submission_id");
+          }
+        });
+      }
     })
   ]);
 };
